+++ src/LegendrePolynomial_t.cu
@@ -36,7 +36,7 @@ __global__
 void SEM::legendre_gauss_nodes_and_weights(int N, deviceFloat* nodes, deviceFloat* weights) {
     const int index = blockIdx.x * blockDim.x + threadIdx.x;
     const int stride = blockDim.x * gridDim.x;
-    const int offset = N * (N + 1) /2;
+    const size_t offset = N * (N + 1) /2;
 
     for (int i = index; i < (N + 1)/2; i += stride) {
         if (N == 1) { // CHECK will enter loop above
